Add tests fo module crypto and correct a bug in decrypt
Bug was in the selection of the generatrix columns, from which the message is extracted.
Now we recompute the generator matrix at the end of the decrypt to find the permutation
giving an independant set of columns. That entails a lot of extra computation (including
the parity check matrix). Solution could be to include the permutation in the secret key.
Or another matrix method could be implemented giving an independant set of columns.