Add enum FieldOption and modify module matrix accordingly. Tests down.
FieldOption denotes the presence or the absence of a field.
In case of an absence of a field, it gives the parameters necessary to generate one.
All matrix constructors use it instead of Rc<F>.
The next step is to implement it for types Polynomial and Goppa and fix the tests.